// packages/domain/src/shared/raw-data-type.ts
export enum RawDataType {
  MARKET_STATS = "MARKET_STATS",
  INDUSTRY_REPORT = "INDUSTRY_REPORT",
  NEWS = "NEWS",
  CUSTOMER_RESEARCH = "CUSTOMER_RESEARCH",
  SNS_ANALYTICS = "SNS_ANALYTICS",
  SEARCH_TRENDS = "SEARCH_TRENDS",
  LOCATION_DATA = "LOCATION_DATA",
  SALES_DATA = "SALES_DATA",
  COMPETITOR_INFO = "COMPETITOR_INFO",
  PARTNER_HEARING = "PARTNER_HEARING",
  FINANCIAL_DATA = "FINANCIAL_DATA",
  EXPERT_HEARING = "EXPERT_HEARING",
}